# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ISP Portal API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List all departments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a department
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{id}:
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update a department
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
        "404":
          description: Not found
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete a department
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /employees:
    get:
      operationId: listEmployees
      tags: [employees]
      summary: List all employees
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of employees
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmployeeSummary"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeSummary"
  /employees/{id}:
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeSummary"
        "404":
          description: Not found
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/UserProfile"
        token:
          type: string
    UserProfile:
      type: object
      required: [id, username, fullName, email, role]
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        isSuperAdmin:
          type: boolean
...
    Role:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        permissions:
          type: array
          items:
            $ref: "#/components/schemas/Permission"
        createdAt:
          type: string
          format: date-time
...
    Permission:
      type: object
      required: [id, name, module, action]
      properties:
        id:
          type: integer
        name:
          type: string
        module:
          type: string
        action:
          type: string
        description:
          type: ["string", "null"]
...
    Department:
      type: object
      required: [id, name, employeeCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        employeeCount:
          type: integer
        createdAt:
          type: string
          format: date-time
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
    EmployeeSummary:
      type: object
      required: [id, employeeId, fullName, hasPortalAccess, createdAt, updatedAt]
      properties:
        id:
          type: integer
        employeeId:
          type: string
        userId:
          type: ["integer", "null"]
        hasPortalAccess:
          type: boolean
        portalUsername:
          type: ["string", "null"]
        fullName:
          type: string
        designation:
          type: ["string", "null"]
        departmentId:
          type: ["integer", "null"]
        departmentName:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
        joiningDate:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        personalContact:
          type: ["string", "null"]
        basicSalary:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    EmployeeInput:
      type: object
      required: [fullName]
      properties:
        fullName:
          type: string
        employeeId:
          type: string
        designation:
          type: string
        departmentId:
          type: integer
        fatherName:
          type: string
        motherName:
          type: string
        dateOfBirth:
          type: string
        joiningDate:
          type: string
        gender:
          type: string
        maritalStatus:
          type: string
        bloodGroup:
          type: string
        nationalId:
          type: string
        basicSalary:
          type: string
        mobileBill:
          type: string
        houseRent:
          type: string
        medical:
          type: string
        food:
          type: string
        otherAllowances:
          type: string
        providentFund:
          type: string
        professionalTax:
          type: string
        incomeTax:
          type: string
        presentAddress:
          type: string
        permanentAddress:
          type: string
        personalContact:
          type: string
        officeContact:
          type: string
        familyContact:
          type: string
        firstReference:
          type: string
        secondReference:
          type: string
        email:
          type: string
        skype:
          type: string
        portalUsername:
          type: string
        portalPassword:
          type: string
        portalRoleId:
          type: integer
        removePortalAccess:
          type: boolean