# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ISP Portal API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeInput {
  fullName: string;
  employeeId?: string;
  designation?: string;
  departmentId?: number;
  fatherName?: string;
  motherName?: string;
  dateOfBirth?: string;
  joiningDate?: string;
  gender?: string;
  maritalStatus?: string;
  bloodGroup?: string;
  nationalId?: string;
  basicSalary?: string;
  mobileBill?: string;
  houseRent?: string;
  medical?: string;
  food?: string;
  otherAllowances?: string;
  providentFund?: string;
  professionalTax?: string;
  incomeTax?: string;
  presentAddress?: string;
  permanentAddress?: string;
  personalContact?: string;
  officeContact?: string;
  familyContact?: string;
  firstReference?: string;
  secondReference?: string;
  email?: string;
  skype?: string;
  portalUsername?: string;
  portalPassword?: string;
  portalRoleId?: number;
  removePortalAccess?: boolean;
}